GuessTheNumber-Manual/src/requirements.txt
boto3
Flask
opentelemetry-distro[otlp]
opentelemetry-sdk-extension-aws
opentelemetry-propagator-aws-xray
